src/requirements.txt
Flask_RESTful==0.3.5
Flask==0.12.1
requests==2.13.0
Mailman==3.1.0b5
PyYAML==3.12
